+++ console/src/main/java/com/elypia/commandler/string/modules/ValidationModule.java
@@ -2,12 +2,9 @@ package com.elypia.commandler.string.modules;
 
 import com.elypia.commandler.annotations.Module;
 import com.elypia.commandler.annotations.*;
-import com.elypia.commandler.annotations.validation.param.*;
-import com.elypia.commandler.parsers.DurationParser;
 import com.elypia.commandler.string.StringHandler;
 
-import java.time.Duration;
-import java.util.concurrent.TimeUnit;
+import javax.validation.constraints.Size;
 
 @Module(name = "Validation", aliases = "valid")
 public class ValidationModule extends StringHandler {
@@ -15,7 +12,7 @@ public class ValidationModule extends StringHandler {
     @Command(name = "Length", aliases = "length")
     @Param(name = "one", help = "The string that's prepended to the next.")
     @Param(name = "two", help = "The string that's appended to the previous.")
-    public String putTogether(@Length(max = 1) String one, @Length(min = 1, max = 10) String two) {
+    public String putTogether(@Size(max = 1) String one, @Size(min = 1, max = 10) String two) {
         return one + two;
     }
 
@@ -26,15 +23,15 @@ public class ValidationModule extends StringHandler {
         return one + two;
     }
 
-    @Command(name = "Option", aliases = "option")
-    @Param(name = "name", help = "Any text of the options: seth or jen.")
-    public String showSelected(@Option({"seth", "jen"}) String name) {
-        return name;
-    }
-
-    @Command(name = "Period", aliases = "period")
-    @Param(name = "time", help = "Any period of time under one day.")
-    public String underADay(@Period(max = 1, unit = TimeUnit.DAYS) Duration time) {
-        return DurationParser.forDisplay(time);
-    }
+//    @Command(name = "Option", aliases = "option")
+//    @Param(name = "name", help = "Any text of the options: seth or jen.")
+//    public String showSelected(@Option({"seth", "jen"}) String name) {
+//        return name;
+//    }
+//
+//    @Command(name = "Period", aliases = "period")
+//    @Param(name = "time", help = "Any period of time under one day.")
+//    public String underADay(@Period(max = 1, unit = TimeUnit.DAYS) Duration time) {
+//        return DurationParser.forDisplay(time);
+//    }
 }