+++ jda/src/main/java/com/elypia/commandler/jda/IJDACommandValidator.java
@@ -1,9 +1,0 @@
-package com.elypia.commandler.jda;
-
-import com.elypia.commandler.impl.ICommandValidator;
-
-import java.lang.annotation.Annotation;
-
-public interface IJDACommandValidator<O extends Annotation> extends ICommandValidator<JDACommand, O> {
-
-}