+++ jda/src/main/java/com/elypia/commandler/jda/IJDAParser.java
@@ -1,7 +1,0 @@
-package com.elypia.commandler.jda;
-
-import com.elypia.commandler.impl.IParser;
-
-public interface IJDAParser<O> extends IParser<JDACommand, O> {
-
-}