+++ jda/src/main/java/com/elypia/commandler/jda/annotations/Emojis.java
@@ -1,9 +1,0 @@
-package com.elypia.commandler.jda.annotations;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Emojis {
-    Emoji[] value();
-}