+++ jda/src/main/java/com/elypia/commandler/jda/annotations/validation/param/Talkable.java
@@ -1,9 +1,0 @@
-package com.elypia.commandler.jda.annotations.validation.param;
-
-import java.lang.annotation.*;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Talkable {
-
-}