+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://gitlab.com/Elypia/Commandler/tree/dev)
 ### Changes
 - Pulled PageBuilder out and created a subproject for called Commandlerdoc, specifically for generating the static webpage.
+- Removed `com.elypia.commandler.jda`, as that will now be a seperate in a seperate repository: `com.elypia.jdacommandler`.